Change translation concept
The concept of translating between languages was changed from a symmetrical one (where words hold references to other words as counterparts) to an asymmetrical one (where a word holds the translation and/or definition as a string directly). This means only one language at a time is now subject to declensions, rhyming, etc, while having static translations into the other language.